# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    post:
      operationId: createReport
      tags: [safety]
      summary: Report a post, video, live stream, comment, story, product, profile or DM
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReportInput" }
      responses:
        "201":
          description: Report filed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "200":
          description: The same person already has an open report on this item
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [already_reported] }
                  id: { type: string, format: uuid }
        "400": { description: "Invalid reason, missing note for \"other\", or self-report" }
        "404": { description: Content not found }
...
  /moderation/reports:
    get:
      operationId: listModerationReports
      tags: [moderation]
      summary: Review queue (moderators only)
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, resolved, all] } }
        - { name: type, in: query, required: false, schema: { $ref: "#/components/schemas/ReportTargetType" } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Queue page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModerationQueue" }
        "403": { description: Moderator access required }
...
components:
  schemas:
...
    ProfileSummary:
      type: object
      required: [userId, name, handle, initials, suspended, deleted]
      properties:
        userId: { type: string }
        name: { type: string }
        handle: { type: string }
        initials: { type: string }
        avatarUrl: { type: ["string", "null"] }
        accountType: { type: ["string", "null"] }
        suspended: { type: boolean }
        deleted: { type: boolean }
...
    ReportTargetType:
      type: string
      enum: [post, video, live, live_comment, comment, story, product, profile, message, community_message, community, review]
    ReportReason:
      type: string
      enum: [spam, harassment, nudity, hate, violence, ip_counterfeit, scam, other]
    CreateReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType: { $ref: "#/components/schemas/ReportTargetType" }
        targetId: { type: string }
        reason: { $ref: "#/components/schemas/ReportReason" }
        note:
          type: string
          maxLength: 1000
          description: Required when reason is "other".
    Report:
      type: object
      required: [id, targetType, targetId, reason, status, source, createdAt]
      properties:
        id: { type: string, format: uuid }
        targetType: { type: string }
        targetId: { type: string }
        targetLabel: { type: ["string", "null"] }
        targetOwnerId: { type: ["string", "null"] }
        contentExcerpt: { type: ["string", "null"] }
        reason: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string, enum: [pending, reviewed, actioned, dismissed] }
        source: { type: string, enum: [user, auto_filter] }
        createdAt: { type: string, format: date-time }
    ModerationQueueItem:
      type: object
      required: [id, status, source, targetType, targetId, reason, createdAt, openReportsOnTarget, ownerPriorActions]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, reviewed, actioned, dismissed] }
        source: { type: string, enum: [user, auto_filter] }
        targetType: { type: string }
        targetId: { type: string }
        targetLabel: { type: ["string", "null"] }
        contentExcerpt: { type: ["string", "null"] }
        reason: { type: string }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        resolution:
          type: ["object", "null"]
          properties:
            action: { type: ["string", "null"] }
            note: { type: ["string", "null"] }
            resolvedAt: { type: ["string", "null"], format: date-time }
        owner:
          oneOf:
            - { $ref: "#/components/schemas/ProfileSummary" }
            - { type: "null" }
        reporter:
          oneOf:
            - { $ref: "#/components/schemas/ProfileSummary" }
            - { type: "null" }
        openReportsOnTarget: { type: integer, minimum: 0 }
        ownerPriorActions: { type: integer, minimum: 0 }
    ModerationQueue:
      type: object
      required: [items, hasMore, summary]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ModerationQueueItem" } }
        hasMore: { type: boolean }
        summary:
          type: object
          required: [open, heldByFilter, resolvedToday]
          properties:
            open: { type: integer }
            heldByFilter: { type: integer }
            resolvedToday: { type: integer }
...